backend/requirements.txt
numpy==1.19.*
Django==3.1.*
asgiref==3.3.0
django-cors-headers==3.6.0
djangorestframework==3.12.1
psycopg2-binary==2.8.6
pytz==2020.4
sqlparse==0.4.1
